Rename TrackerUtils to distinguish it
